Update Install VirtualBox.sh

+++ Library/Install VirtualBox.sh
@@ -14,13 +14,17 @@ done
 version=$(curl --silent --insecure "https://download.virtualbox.org/virtualbox/LATEST.TXT")
 echo "==> Install 'VirtualBox ${version} Oracle VM VirtualBox Extension Pack'"
 
-curl \
-  --silent \
-  --output "/private/tmp/Oracle_VM_VirtualBox_Extension_Pack-${version}.vbox-extpack" \
-  "https://download.virtualbox.org/virtualbox/${version}/Oracle_VM_VirtualBox_Extension_Pack-${version}.vbox-extpack"
+if [ "$(/usr/local/bin/aky --cat virtualbox | jq -r '.recipe.version')" = "${version}" ]; then
 
-echo "y" | /usr/local/bin/VBoxManage extpack install --replace "/private/tmp/Oracle_VM_VirtualBox_Extension_Pack-${version}.vbox-extpack"
+  curl \
+    --silent \
+    --output "/private/tmp/Oracle_VM_VirtualBox_Extension_Pack-${version}.vbox-extpack" \
+    "https://download.virtualbox.org/virtualbox/${version}/Oracle_VM_VirtualBox_Extension_Pack-${version}.vbox-extpack"
 
-if [ -s "/private/tmp/Oracle_VM_VirtualBox_Extension_Pack-${version}.vbox-extpack" ]; then
-  rm -rf "/private/tmp/Oracle_VM_VirtualBox_Extension_Pack-${version}.vbox-extpack"
+  echo "y" | /usr/local/bin/VBoxManage extpack install --replace "/private/tmp/Oracle_VM_VirtualBox_Extension_Pack-${version}.vbox-extpack"
+
+  if [ -s "/private/tmp/Oracle_VM_VirtualBox_Extension_Pack-${version}.vbox-extpack" ]; then
+    rm -rf "/private/tmp/Oracle_VM_VirtualBox_Extension_Pack-${version}.vbox-extpack"
+  fi
+
 fi
